requirements.txt
Django>=5.2.3
djangorestframework>=3.14.0
pika>=1.3.0
